+++ @xen-orchestra/lite/src/modules/network/jobs/network-delete-args.ts
@@ -1,0 +1,7 @@
+import type { XenApiNetwork } from '@/libs/xen-api/xen-api.types.ts'
+import { defineJobArg } from '@core/packages/job'
+
+export const networksArg = defineJobArg({
+  identify: (network: XenApiNetwork) => network.$ref,
+  toArray: true,
+})