+++ @xen-orchestra/web-core/lib/components/menu/VtsActionsMenu.vue
@@ -16,6 +16,7 @@
       <MenuItem
         v-for="(action, index) of actions"
         :key="index"
+        :class="{ danger: action.danger }"
         :icon="action.icon"
         :disabled="action.disabled"
         :busy="action.busy"
@@ -67,6 +68,7 @@ type BaseActionItem = {
   icon?: IconName
   disabled?: boolean
   busy?: boolean
+  danger?: boolean
 }
 
 export type LeafActionItem = BaseActionItem & {
@@ -85,3 +87,9 @@ function isGroupAction(action: ActionItem): action is GroupActionItem {
   return 'children' in action
 }
 </script>
+
+<style lang="postcss" scoped>
+.danger {
+  color: var(--color-danger-item-base);
+}
+</style>